Небольшая уборка
- Переименовано всякое
- Добавлен централизованный ServiceCollectionExtensions
